rag/requirements.txt
langchain>=0.3
langchain-community>=0.3
langchain-ollama>=1.0
langgraph>=0.2
faiss-cpu>=1.8
rank-bm25>=0.2
jsonschema>=4.20
pydantic>=2.5
fastapi>=0.110
uvicorn>=0.27
python-dotenv>=1.0
sentence-transformers>=3.0
torch>=2.0
